refactor: consolidate sign flow into sign-utils.ts
Move SignFlowAdapter, SignFlowResult, and executeSignFlow from the
standalone sign-flow.ts into the existing sign-utils.ts module.
Removes the extra file while keeping all functionality co-located
with the other sign utilities.

Copilot-Session: 12624e77-c374-4b9f-a6d3-3c44f91327cb
