+++ src/sign-utils.ts
@@ -55,3 +55,82 @@ export async function findWorkspaceArtifacts(
 	withStats.sort((a, b) => b.mtime - a.mtime);
 	return withStats.map((s) => s.path);
 }
+
+// ──────────────────────────────────────────────────────
+// Sign flow — adapter-based orchestration for testability.
+// The VS Code-facing wiring (QuickPick, file dialogs, terminal) remains in
+// extension.ts; this section owns only the decision logic and delegates UI
+// operations through an injectable adapter interface.
+// ──────────────────────────────────────────────────────
+
+export interface SignFlowAdapter {
+	/** Show a QuickPick to let the user choose a signable artifact. */
+	pickSignableFile(workspacePath: string): Promise<string | undefined>;
+
+	/** Show a QuickPick to let the user choose a signing certificate. */
+	pickCertificateFile(workspacePath: string): Promise<string | undefined>;
+
+	/** Execute the sign CLI command. */
+	runSignCommand(extensionPath: string, command: string, workspacePath: string): Promise<void>;
+}
+
+export interface SignFlowResult {
+	/** Whether the file picker was shown (false when prefilled path provided). */
+	filePickerShown: boolean;
+	/** Whether the certificate picker was shown. */
+	certPickerShown: boolean;
+	/** The CLI command string that was executed, if any. */
+	commandExecuted: string | undefined;
+	/** The file path that was signed, if any. */
+	filePath: string | undefined;
+	/** The certificate path used, if any. */
+	certPath: string | undefined;
+}
+
+/**
+ * Run the sign-package flow.
+ *
+ * When `prefilledFilePath` is provided (e.g. from the post-pack "Sign" action),
+ * the file picker is skipped and the flow proceeds directly to the certificate
+ * picker. When omitted, the full QuickPick discovery flow is used.
+ *
+ * @returns A result object describing what happened during the flow.
+ */
+export async function executeSignFlow(
+	adapter: SignFlowAdapter,
+	extensionPath: string,
+	workspacePath: string,
+	prefilledFilePath?: string
+): Promise<SignFlowResult> {
+	const result: SignFlowResult = {
+		filePickerShown: false,
+		certPickerShown: false,
+		commandExecuted: undefined,
+		filePath: undefined,
+		certPath: undefined
+	};
+
+	let filePath = prefilledFilePath;
+	if (!filePath) {
+		result.filePickerShown = true;
+		filePath = await adapter.pickSignableFile(workspacePath);
+	}
+
+	if (!filePath) {
+		return result;
+	}
+	result.filePath = filePath;
+
+	result.certPickerShown = true;
+	const certPath = await adapter.pickCertificateFile(workspacePath);
+	if (!certPath) {
+		return result;
+	}
+	result.certPath = certPath;
+
+	const command = buildSignCommand(filePath, certPath);
+	result.commandExecuted = command;
+	await adapter.runSignCommand(extensionPath, command, workspacePath);
+
+	return result;
+}
